02-coding-interview/backend/requirements.txt
fastapi==0.123.5
uvicorn==0.38.0
websockets==15.0.1
python-socketio==5.15.0
